+++ src/Humanizer/Bytes/ByteRate.cs
@@ -5,25 +5,56 @@ using System.Text;
 
 namespace Humanizer.Bytes
 {
+
+    /// <summary>
+    /// Class to hold a ByteSize and a measurement interval, for the purpose of calculating the rate of transfer
+    /// </summary>
     public class ByteRate
     {
-        public ByteSize size { get; private set;}
+        /// <summary>
+        /// Quantity of bytes
+        /// </summary>
+        /// <returns></returns>
+        public ByteSize Size { get; private set;}
 
-        public TimeSpan interval { get; private set; }
+        /// <summary>
+        /// Interval that bytes were transferred in
+        /// </summary>
+        /// <returns></returns>
+        public TimeSpan Interval { get; private set; }
 
+        /// <summary>
+        /// Create a ByteRate with given quantity of bytes across an interval
+        /// </summary>
+        /// <param name="size"></param>
+        /// <param name="interval"></param>
         public ByteRate(ByteSize size, TimeSpan interval)
         {
-            this.size = size;
-            this.interval = interval;
+            this.Size = size;
+            this.Interval = interval;
         }
 
         /// <summary>
-        /// Return this rate as a string, EG "2 MB/s"
+        /// Calculate rate for the quantity of bytes and interval defined by this instance
         /// </summary>
+        /// <param name="timeUnit">Unit of time to calculate rate for (defaults is per second)</param>
         /// <returns></returns>
-        public string ToRatePerSecond()
+        public string Humanize(TimeUnit timeUnit = TimeUnit.Second)
         {
-            return (new ByteSize(size.Bytes / interval.TotalSeconds)).Humanize() + "/s";
+            var displayInterval = TimeSpan.FromSeconds(1);
+            var displayUnit = "s";
+
+            if (timeUnit == TimeUnit.Minute)
+            {
+                displayInterval = TimeSpan.FromMinutes(1);
+                displayUnit = "min";
+            } else if(timeUnit == TimeUnit.Hour)
+            {
+                displayInterval = TimeSpan.FromHours(1);
+                displayUnit = "hour";
+            }
+
+            return (new ByteSize(Size.Bytes / Interval.TotalSeconds * displayInterval.TotalSeconds)).Humanize() + '/' + displayUnit;
         }
     }
 }