+++ src/Humanizer/Bytes/TimeUnit.cs
@@ -1,0 +1,23 @@
+﻿namespace Humanizer.Bytes
+{
+    /// <summary>
+    /// Units of time for which byte rates can be calculated
+    /// </summary>
+    public enum TimeUnit
+    {
+        /// <summary>
+        /// One hour
+        /// </summary>
+        Hour,
+
+        /// <summary>
+        /// One minute
+        /// </summary>
+        Minute,
+
+        /// <summary>
+        /// One second
+        /// </summary>
+        Second
+    }
+}